pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testscout"
version = "0.1.0"
description = "AI-Powered E2E Testing Framework - Test web apps with natural language using AI vision"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "testScout Contributors", email = "contact@testscout.dev"}
]
maintainers = [
    {name = "testScout Contributors", email = "contact@testscout.dev"}
]
keywords = [
    "testing",
    "e2e",
    "end-to-end",
    "playwright",
    "ai",
    "vision",
    "automation",
    "gemini",
    "openai",
    "set-of-marks",
    "visual-testing",
    "qa",
    "autonomous-testing",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "playwright>=1.40.0",
]

[project.optional-dependencies]
gemini = [
    "google-generativeai>=0.3.0",
]
openai = [
    "openai>=1.0.0",
]
all = [
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-playwright>=0.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
testscout = "testscout.cli:main"

[project.urls]
Homepage = "https://github.com/rhowardstone/testScout"
Repository = "https://github.com/rhowardstone/testScout"
Issues = "https://github.com/rhowardstone/testScout/issues"
Discussions = "https://github.com/rhowardstone/testScout/discussions"
Documentation = "https://github.com/rhowardstone/testScout/tree/main/docs"
Changelog = "https://github.com/rhowardstone/testScout/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]
include = ["testscout*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
testscout = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "ai_e2e: marks tests as AI-powered E2E tests (may be skipped without API key)",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501", "E722"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
